+++ udata/tests/api/test_auth_api.py
@@ -90,7 +90,7 @@ class APIAuthTest:
         assert response.json == {'success': True}
 
     def test_oauth_auth(self, api, oauth):
-        '''Should handle  OAuth header authentication'''
+        '''Should handle OAuth header authentication'''
         user = UserFactory()
         token = OAuth2Token.objects.create(
             client=oauth,