+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Current (in progress)
 
+- Add activity.key filter to activity.atom feed [#2293](https://github.com/opendatateam/udata/pull/2293)
+- Allow `Authorization` as CORS header and OAuth minor fixes [#2298](https://github.com/opendatateam/udata/pull/2298)
+
+## 1.6.14 (2019-08-14)
+
 - Cleanup `permitted_reuses` data (migration) [#2244](https://github.com/opendatateam/udata/pull/2244)
 - Proper form errors handling on nested fields [#2246](https://github.com/opendatateam/udata/pull/2246)
 - JS models load/save/update consistency (`loading` always `true` on query, always handle error, no more silent errors) [#2247](https://github.com/opendatateam/udata/pull/2247)